requirements.txt
inquirer
python-dotenv
pydantic
requests
rich
